Move test_multiindex.py
