+++ python/cudf/cudf/tests/dataframe/indexing/test_iloc.py
@@ -1,0 +1,173 @@
+# Copyright (c) 2025, NVIDIA CORPORATION.
+
+
+import cupy as cp
+import numpy as np
+import pandas as pd
+import pytest
+
+import cudf
+from cudf.testing import assert_eq
+
+
+@pytest.mark.parametrize(
+    "iloc_rows",
+    [
+        0,
+        1,
+        slice(None, 0),
+        slice(None, 1),
+        slice(0, 1),
+        slice(1, 2),
+        slice(0, 2),
+        slice(0, None),
+        slice(1, None),
+    ],
+)
+@pytest.mark.parametrize(
+    "iloc_columns",
+    [
+        0,
+        1,
+        slice(None, 0),
+        slice(None, 1),
+        slice(0, 1),
+        slice(1, 2),
+        slice(0, 2),
+        slice(0, None),
+        slice(1, None),
+    ],
+)
+def test_multiindex_iloc(iloc_rows, iloc_columns):
+    rng = np.random.default_rng(seed=0)
+    pdf = pd.DataFrame(rng.random(size=(7, 5)))
+    gdf = cudf.from_pandas(pdf)
+    pdfIndex = pd.MultiIndex(
+        [
+            ["a", "b", "c"],
+            ["house", "store", "forest"],
+            ["clouds", "clear", "storm"],
+            ["fire", "smoke", "clear"],
+            [
+                np.datetime64("2001-01-01", "ns"),
+                np.datetime64("2002-01-01", "ns"),
+                np.datetime64("2003-01-01", "ns"),
+            ],
+        ],
+        [
+            [0, 0, 0, 0, 1, 1, 2],
+            [1, 1, 1, 1, 0, 0, 2],
+            [0, 0, 2, 2, 2, 0, 1],
+            [0, 0, 0, 1, 2, 0, 1],
+            [1, 0, 1, 2, 0, 0, 1],
+        ],
+    )
+    pdfIndex.names = ["alpha", "location", "weather", "sign", "timestamp"]
+    gdfIndex = cudf.from_pandas(pdfIndex)
+    assert_eq(pdfIndex, gdfIndex)
+    pdf.index = pdfIndex
+    gdf.index = gdfIndex
+    presult = pdf.iloc[iloc_rows, iloc_columns]
+    gresult = gdf.iloc[iloc_rows, iloc_columns]
+    if isinstance(gresult, cudf.DataFrame):
+        assert_eq(
+            presult, gresult, check_index_type=False, check_column_type=False
+        )
+    else:
+        assert_eq(presult, gresult, check_index_type=False, check_dtype=False)
+
+
+def test_multiindex_iloc_scalar():
+    arrays = [["a", "a", "b", "b"], [1, 2, 3, 4]]
+    tuples = list(zip(*arrays, strict=True))
+    idx = cudf.MultiIndex.from_tuples(tuples)
+    gdf = cudf.DataFrame(
+        {"first": cp.random.rand(4), "second": cp.random.rand(4)}
+    )
+    gdf.index = idx
+
+    pdf = gdf.to_pandas()
+    assert_eq(pdf.iloc[3], gdf.iloc[3])
+
+
+@pytest.mark.parametrize(
+    "iloc_rows",
+    [
+        0,
+        1,
+        slice(None, 0),
+        slice(None, 1),
+        slice(0, 1),
+        slice(1, 2),
+        slice(0, 2),
+        slice(0, None),
+        slice(1, None),
+    ],
+)
+@pytest.mark.parametrize(
+    "iloc_columns",
+    [
+        0,
+        1,
+        slice(None, 0),
+        slice(None, 1),
+        slice(0, 1),
+        slice(1, 2),
+        slice(0, 2),
+        slice(0, None),
+        slice(1, None),
+    ],
+)
+def test_multicolumn_iloc(iloc_rows, iloc_columns):
+    rng = np.random.default_rng(seed=0)
+    pdf = pd.DataFrame(rng.random(size=(7, 5)))
+    gdf = cudf.from_pandas(pdf)
+    pdfIndex = pd.MultiIndex(
+        [
+            ["a", "b", "c"],
+            ["house", "store", "forest"],
+            ["clouds", "clear", "storm"],
+            ["fire", "smoke", "clear"],
+            [
+                np.datetime64("2001-01-01", "ns"),
+                np.datetime64("2002-01-01", "ns"),
+                np.datetime64("2003-01-01", "ns"),
+            ],
+        ],
+        [
+            [0, 0, 0, 0, 1, 1, 2],
+            [1, 1, 1, 1, 0, 0, 2],
+            [0, 0, 2, 2, 2, 0, 1],
+            [0, 0, 0, 1, 2, 0, 1],
+            [1, 0, 1, 2, 0, 0, 1],
+        ],
+    )
+    pdfIndex.names = ["alpha", "location", "weather", "sign", "timestamp"]
+    gdfIndex = cudf.from_pandas(pdfIndex)
+    assert_eq(pdfIndex, gdfIndex)
+    pdf.index = pdfIndex
+    gdf.index = gdfIndex
+    pdf = pdf.T
+    gdf = gdf.T
+    presult = pdf.iloc[iloc_rows, iloc_columns]
+    gresult = gdf.iloc[iloc_rows, iloc_columns]
+    if hasattr(gresult, "name") and isinstance(gresult.name, tuple):
+        name = gresult.name[len(gresult.name) - 1]
+        if isinstance(name, str) and "cudf" in name:
+            gresult.name = name
+    if isinstance(presult, pd.DataFrame):
+        assert_eq(
+            presult, gresult, check_index_type=False, check_column_type=False
+        )
+    else:
+        assert_eq(presult, gresult, check_index_type=False, check_dtype=False)
+
+
+def test_multiindex_multicolumn_zero_row_slice():
+    gdf = cudf.DataFrame(
+        {"x": [1, 5, 3, 4, 1], "y": [1, 1, 2, 2, 5], "z": [1, 2, 3, 4, 5]}
+    )
+    pdf = gdf.to_pandas()
+    gdg = gdf.groupby(["x", "y"]).agg({"z": ["count"]}).iloc[:0]
+    pdg = pdf.groupby(["x", "y"]).agg({"z": ["count"]}).iloc[:0]
+    assert_eq(pdg, gdg, check_dtype=False)
