+++ python/cudf/cudf/tests/dataframe/indexing/test_loc.py
@@ -1,5 +1,6 @@
 # Copyright (c) 2025, NVIDIA CORPORATION.
 import re
+from contextlib import contextmanager
 
 import numpy as np
 import pandas as pd
@@ -7,6 +8,7 @@ import pytest
 
 import cudf
 from cudf.testing import assert_eq
+from cudf.testing._utils import assert_exceptions_equal, expect_warning_if
 
 
 def test_dataframe_midx_columns_loc():
@@ -171,3 +173,425 @@ def test_loc_setitem_add_column_partial_12801():
     cdf.loc[cdf.a < 2, "b"] = 1
 
     assert_eq(df, cdf)
+
+
+@contextmanager
+def expect_pandas_performance_warning(idx):
+    with expect_warning_if(
+        (not isinstance(idx[0], tuple) and len(idx) > 2)
+        or (isinstance(idx[0], tuple) and len(idx[0]) > 2),
+        pd.errors.PerformanceWarning,
+    ):
+        yield
+
+
+@pytest.mark.parametrize(
+    "key_tuple",
+    [
+        # return 2 rows, 0 remaining keys = dataframe with entire index
+        ("a", "store", "clouds", "fire"),
+        (("a", "store", "clouds", "fire"), slice(None)),
+        # return 2 rows, 1 remaining key = dataframe with n-k index columns
+        ("a", "store", "storm"),
+        (("a", "store", "storm"), slice(None)),
+        # return 2 rows, 2 remaining keys = dataframe with n-k index columns
+        ("a", "store"),
+        (("a", "store"), slice(None)),
+        # return 2 rows, n-1 remaining keys = dataframe with n-k index columns
+        ("a",),
+        "a",
+        "b",
+        "c",
+        (("a",), slice(None)),
+        # return 1 row, 0 remaining keys = dataframe with entire index
+        ("a", "store", "storm", "smoke"),
+        (("a", "store", "storm", "smoke"), slice(None)),
+        # return 1 row and 1 remaining key = series
+        ("c", "forest", "clear"),
+        (("c", "forest", "clear"), slice(None)),
+    ],
+)
+def test_multiindex_loc(key_tuple):
+    rng = np.random.default_rng(seed=0)
+    pdf = pd.DataFrame(rng.random(size=(7, 5)))
+    pdfIndex = pd.MultiIndex(
+        [
+            ["a", "b", "c"],
+            ["house", "store", "forest"],
+            ["clouds", "clear", "storm"],
+            ["fire", "smoke", "clear"],
+            [
+                np.datetime64("2001-01-01", "ns"),
+                np.datetime64("2002-01-01", "ns"),
+                np.datetime64("2003-01-01", "ns"),
+            ],
+        ],
+        [
+            [0, 0, 0, 0, 1, 1, 2],
+            [1, 1, 1, 1, 0, 0, 2],
+            [0, 0, 2, 2, 2, 0, 1],
+            [0, 0, 0, 1, 2, 0, 1],
+            [1, 0, 1, 2, 0, 0, 1],
+        ],
+    )
+    pdfIndex.names = ["alpha", "location", "weather", "sign", "timestamp"]
+    gdfIndex = cudf.from_pandas(pdfIndex)
+    assert_eq(pdfIndex, gdfIndex)
+    gdf = cudf.from_pandas(pdf)
+    pdf.index = pdfIndex
+    gdf.index = gdfIndex
+    # The index is unsorted, which makes things slow but is fine for testing.
+    with expect_pandas_performance_warning(key_tuple):
+        expected = pdf.loc[key_tuple]
+    got = gdf.loc[key_tuple].sort_index()
+    assert_eq(expected.sort_index(), got)
+
+    with cudf.option_context("mode.pandas_compatible", True):
+        got = gdf.loc[key_tuple]
+    assert_eq(expected, got)
+
+
+@pytest.mark.parametrize("second_val", [[0, 1], [1, 0]])
+def test_multiindex_compatible_ordering(second_val):
+    indexer = (([1, 1], second_val), slice(None))
+    df = pd.DataFrame(
+        {"a": [1, 1, 2, 3], "b": [1, 0, 1, 1], "c": [1, 2, 3, 4]}
+    ).set_index(["a", "b"])
+    cdf = cudf.from_pandas(df)
+    expect = df.loc[indexer]
+    with cudf.option_context("mode.pandas_compatible", True):
+        actual = cdf.loc[indexer]
+    assert_eq(actual, expect)
+
+
+@pytest.mark.parametrize(
+    "arg",
+    [
+        slice(("a", "store"), ("b", "house")),
+        slice(None, ("b", "house")),
+        slice(("a", "store"), None),
+        slice(None),
+    ],
+)
+def test_multiindex_loc_slice(arg):
+    rng = np.random.default_rng(seed=0)
+    pdf = pd.DataFrame(rng.random(size=(7, 5)))
+    pdfIndex = pd.MultiIndex(
+        [
+            ["a", "b", "c"],
+            ["house", "store", "forest"],
+            ["clouds", "clear", "storm"],
+            ["fire", "smoke", "clear"],
+            [
+                np.datetime64("2001-01-01", "ns"),
+                np.datetime64("2002-01-01", "ns"),
+                np.datetime64("2003-01-01", "ns"),
+            ],
+        ],
+        [
+            [0, 0, 0, 0, 1, 1, 2],
+            [1, 1, 1, 1, 0, 0, 2],
+            [0, 0, 2, 2, 2, 0, 1],
+            [0, 0, 0, 1, 2, 0, 1],
+            [1, 0, 1, 2, 0, 0, 1],
+        ],
+    )
+    pdfIndex.names = ["alpha", "location", "weather", "sign", "timestamp"]
+    gdf = cudf.from_pandas(pdf)
+    gdfIndex = cudf.from_pandas(pdfIndex)
+    pdf = pdf.copy(deep=False)
+    pdf.index = pdfIndex
+    gdf.index = gdfIndex
+    assert_eq(pdf.loc[arg], gdf.loc[arg])
+
+
+def test_multiindex_loc_errors():
+    rng = np.random.default_rng(seed=0)
+    pdf = pd.DataFrame(rng.random(size=(7, 5)))
+    pdfIndex = pd.MultiIndex(
+        [
+            ["a", "b", "c"],
+            ["house", "store", "forest"],
+            ["clouds", "clear", "storm"],
+            ["fire", "smoke", "clear"],
+            [
+                np.datetime64("2001-01-01", "ns"),
+                np.datetime64("2002-01-01", "ns"),
+                np.datetime64("2003-01-01", "ns"),
+            ],
+        ],
+        [
+            [0, 0, 0, 0, 1, 1, 2],
+            [1, 1, 1, 1, 0, 0, 2],
+            [0, 0, 2, 2, 2, 0, 1],
+            [0, 0, 0, 1, 2, 0, 1],
+            [1, 0, 1, 2, 0, 0, 1],
+        ],
+    )
+    pdfIndex.names = ["alpha", "location", "weather", "sign", "timestamp"]
+    gdfIndex = cudf.from_pandas(pdfIndex)
+    gdf = cudf.from_pandas(pdf)
+    gdf.index = gdfIndex
+
+    with pytest.raises(KeyError):
+        gdf.loc[("a", "store", "clouds", "foo")]
+    with pytest.raises(IndexError):
+        gdf.loc[
+            ("a", "store", "clouds", "fire", "x", "y")
+        ]  # too many indexers
+    with pytest.raises(IndexError):
+        gdf.loc[slice(None, ("a", "store", "clouds", "fire", "x", "y"))]
+
+
+def test_multiindex_loc_then_column():
+    rng = np.random.default_rng(seed=0)
+    pdf = pd.DataFrame(rng.random(size=(7, 5)))
+    pdfIndex = pd.MultiIndex(
+        [
+            ["a", "b", "c"],
+            ["house", "store", "forest"],
+            ["clouds", "clear", "storm"],
+            ["fire", "smoke", "clear"],
+            [
+                np.datetime64("2001-01-01", "ns"),
+                np.datetime64("2002-01-01", "ns"),
+                np.datetime64("2003-01-01", "ns"),
+            ],
+        ],
+        [
+            [0, 0, 0, 0, 1, 1, 2],
+            [1, 1, 1, 1, 0, 0, 2],
+            [0, 0, 2, 2, 2, 0, 1],
+            [0, 0, 0, 1, 2, 0, 1],
+            [1, 0, 1, 2, 0, 0, 1],
+        ],
+    )
+    pdfIndex.names = ["alpha", "location", "weather", "sign", "timestamp"]
+    gdf = cudf.from_pandas(pdf)
+    gdfIndex = cudf.from_pandas(pdfIndex)
+    assert_eq(pdfIndex, gdfIndex)
+    pdf.index = pdfIndex
+    gdf.index = gdfIndex
+    # The index is unsorted, which makes things slow but is fine for testing.
+    with pytest.warns(pd.errors.PerformanceWarning):
+        expected = pdf.loc[("a", "store", "clouds", "fire"), :][0]
+    got = gdf.loc[("a", "store", "clouds", "fire"), :][0]
+    assert_eq(expected, got)
+
+
+def test_multiindex_loc_rows_0():
+    rng = np.random.default_rng(seed=0)
+    pdf = pd.DataFrame(rng.random(size=(7, 5)))
+    pdfIndex = pd.MultiIndex(
+        [
+            ["a", "b", "c"],
+            ["house", "store", "forest"],
+            ["clouds", "clear", "storm"],
+            ["fire", "smoke", "clear"],
+            [
+                np.datetime64("2001-01-01", "ns"),
+                np.datetime64("2002-01-01", "ns"),
+                np.datetime64("2003-01-01", "ns"),
+            ],
+        ],
+        [
+            [0, 0, 0, 0, 1, 1, 2],
+            [1, 1, 1, 1, 0, 0, 2],
+            [0, 0, 2, 2, 2, 0, 1],
+            [0, 0, 0, 1, 2, 0, 1],
+            [1, 0, 1, 2, 0, 0, 1],
+        ],
+    )
+    pdfIndex.names = ["alpha", "location", "weather", "sign", "timestamp"]
+    gdfIndex = cudf.from_pandas(pdfIndex)
+    gdf = cudf.from_pandas(pdf)
+    pdf.index = pdfIndex
+    gdf.index = gdfIndex
+
+    assert_exceptions_equal(
+        lfunc=pdf.loc.__getitem__,
+        rfunc=gdf.loc.__getitem__,
+        lfunc_args_and_kwargs=([(("d",), slice(None, None, None))],),
+        rfunc_args_and_kwargs=([(("d",), slice(None, None, None))],),
+    )
+
+
+def test_multiindex_loc_rows_1_2_key():
+    rng = np.random.default_rng(seed=0)
+    pdf = pd.DataFrame(rng.random(size=(7, 5)))
+    pdfIndex = pd.MultiIndex(
+        [
+            ["a", "b", "c"],
+            ["house", "store", "forest"],
+            ["clouds", "clear", "storm"],
+            ["fire", "smoke", "clear"],
+            [
+                np.datetime64("2001-01-01", "ns"),
+                np.datetime64("2002-01-01", "ns"),
+                np.datetime64("2003-01-01", "ns"),
+            ],
+        ],
+        [
+            [0, 0, 0, 0, 1, 1, 2],
+            [1, 1, 1, 1, 0, 0, 2],
+            [0, 0, 2, 2, 2, 0, 1],
+            [0, 0, 0, 1, 2, 0, 1],
+            [1, 0, 1, 2, 0, 0, 1],
+        ],
+    )
+    pdfIndex.names = ["alpha", "location", "weather", "sign", "timestamp"]
+    gdfIndex = cudf.from_pandas(pdfIndex)
+    gdf = cudf.from_pandas(pdf)
+    pdf.index = pdfIndex
+    gdf.index = gdfIndex
+    assert_eq(pdf.loc[("c", "forest"), :], gdf.loc[("c", "forest"), :])
+
+
+def test_multiindex_loc_rows_1_1_key():
+    rng = np.random.default_rng(seed=0)
+    pdf = pd.DataFrame(rng.random(size=(7, 5)))
+    pdfIndex = pd.MultiIndex(
+        [
+            ["a", "b", "c"],
+            ["house", "store", "forest"],
+            ["clouds", "clear", "storm"],
+            ["fire", "smoke", "clear"],
+            [
+                np.datetime64("2001-01-01", "ns"),
+                np.datetime64("2002-01-01", "ns"),
+                np.datetime64("2003-01-01", "ns"),
+            ],
+        ],
+        [
+            [0, 0, 0, 0, 1, 1, 2],
+            [1, 1, 1, 1, 0, 0, 2],
+            [0, 0, 2, 2, 2, 0, 1],
+            [0, 0, 0, 1, 2, 0, 1],
+            [1, 0, 1, 2, 0, 0, 1],
+        ],
+    )
+    pdfIndex.names = ["alpha", "location", "weather", "sign", "timestamp"]
+    gdfIndex = cudf.from_pandas(pdfIndex)
+    gdf = cudf.from_pandas(pdf)
+    pdf.index = pdfIndex
+    gdf.index = gdfIndex
+    assert_eq(pdf.loc[("c",), :], gdf.loc[("c",), :])
+
+
+def test_multiindex_index_single_row():
+    arrays = [["a", "a", "b", "b"], [1, 2, 3, 4]]
+    tuples = list(zip(*arrays, strict=True))
+    idx = cudf.MultiIndex.from_tuples(tuples)
+    gdf = cudf.DataFrame({"first": range(4), "second": range(4)})
+    gdf.index = idx
+    pdf = gdf.to_pandas()
+    assert_eq(pdf.loc[("b", 3)], gdf.loc[("b", 3)])
+
+
+@pytest.mark.parametrize("idx_get", [(0, 0), (0, 1), (1, 0), (1, 1)])
+@pytest.mark.parametrize("cols_get", [0, 1, [0, 1], [1, 0], [1], [0]])
+def test_multiindex_loc_scalar(idx_get, cols_get):
+    idx = cudf.MultiIndex.from_tuples([(0, 0), (0, 1), (1, 0), (1, 1)])
+    df = cudf.DataFrame({0: range(4), 1: range(10, 50, 10)}, index=idx)
+    pdf = df.to_pandas()
+
+    actual = df.loc[idx_get, cols_get]
+    expected = pdf.loc[idx_get, cols_get]
+
+    assert_eq(actual, expected)
+
+
+def test_multiindex_rows_with_wildcard():
+    rng = np.random.default_rng(seed=0)
+    pdf = pd.DataFrame(rng.random(size=(7, 5)))
+    gdf = cudf.from_pandas(pdf)
+    pdfIndex = pd.MultiIndex(
+        [
+            ["a", "b", "c"],
+            ["house", "store", "forest"],
+            ["clouds", "clear", "storm"],
+            ["fire", "smoke", "clear"],
+            [
+                np.datetime64("2001-01-01", "ns"),
+                np.datetime64("2002-01-01", "ns"),
+                np.datetime64("2003-01-01", "ns"),
+            ],
+        ],
+        [
+            [0, 0, 0, 0, 1, 1, 2],
+            [1, 1, 1, 1, 0, 0, 2],
+            [0, 0, 2, 2, 2, 0, 1],
+            [0, 0, 0, 1, 2, 0, 1],
+            [1, 0, 1, 2, 0, 0, 1],
+        ],
+    )
+    pdfIndex.names = ["alpha", "location", "weather", "sign", "timestamp"]
+    gdfIndex = cudf.from_pandas(pdfIndex)
+    pdf.index = pdfIndex
+    gdf.index = gdfIndex
+    # The index is unsorted, which makes things slow but is fine for testing.
+    with pytest.warns(pd.errors.PerformanceWarning):
+        assert_eq(
+            pdf.loc[("a",), :].sort_index(), gdf.loc[("a",), :].sort_index()
+        )
+        assert_eq(
+            pdf.loc[(("a"), ("store")), :].sort_index(),
+            gdf.loc[(("a"), ("store")), :].sort_index(),
+        )
+        assert_eq(
+            pdf.loc[(("a"), ("store"), ("storm")), :].sort_index(),
+            gdf.loc[(("a"), ("store"), ("storm")), :].sort_index(),
+        )
+        assert_eq(
+            pdf.loc[(("a"), ("store"), ("storm"), ("smoke")), :].sort_index(),
+            gdf.loc[(("a"), ("store"), ("storm"), ("smoke")), :].sort_index(),
+        )
+        assert_eq(
+            pdf.loc[(slice(None), "store"), :].sort_index(),
+            gdf.loc[(slice(None), "store"), :].sort_index(),
+        )
+        assert_eq(
+            pdf.loc[(slice(None), slice(None), "storm"), :].sort_index(),
+            gdf.loc[(slice(None), slice(None), "storm"), :].sort_index(),
+        )
+        assert_eq(
+            pdf.loc[
+                (slice(None), slice(None), slice(None), "smoke"), :
+            ].sort_index(),
+            gdf.loc[
+                (slice(None), slice(None), slice(None), "smoke"), :
+            ].sort_index(),
+        )
+
+
+def test_multicolumn_loc():
+    rng = np.random.default_rng(seed=0)
+    pdf = pd.DataFrame(rng.random(size=(7, 5)))
+    pdfIndex = pd.MultiIndex(
+        [
+            ["a", "b", "c"],
+            ["house", "store", "forest"],
+            ["clouds", "clear", "storm"],
+            ["fire", "smoke", "clear"],
+            [
+                np.datetime64("2001-01-01", "ns"),
+                np.datetime64("2002-01-01", "ns"),
+                np.datetime64("2003-01-01", "ns"),
+            ],
+        ],
+        [
+            [0, 0, 0, 0, 1, 1, 2],
+            [1, 1, 1, 1, 0, 0, 2],
+            [0, 0, 2, 2, 2, 0, 1],
+            [0, 0, 0, 1, 2, 0, 1],
+            [1, 0, 1, 2, 0, 0, 1],
+        ],
+    )
+    pdfIndex.names = ["alpha", "location", "weather", "sign", "timestamp"]
+    pdf = pdf.T
+    pdf.columns = pdfIndex
+    gdf = cudf.from_pandas(pdf)
+    assert_eq(pdf.loc[:, "a"], gdf.loc[:, "a"])
+    assert_eq(pdf.loc[:, ("a", "store")], gdf.loc[:, ("a", "store")])
+    assert_eq(pdf.loc[:, "a":"b"], gdf.loc[:, "a":"b"])
+    assert_eq(pdf.loc[:, ["a", "b"]], gdf.loc[:, ["a", "b"]])
